fix a bunch of issues with chromatic tests
