+++ apps/hackathon/src/utils/anchor-date.ts
@@ -1,0 +1,2 @@
+export const anchorDate = new Date("2025-06-29");
+export const anchorTimestamp = anchorDate.getTime();
