+++ libs/chromatic/tsconfig.storybook.json
@@ -23,7 +23,8 @@
     "src/**/*.stories.tsx",
     "src/**/*.stories.mdx",
     ".storybook/*.js",
-    ".storybook/*.ts"
+    ".storybook/*.ts",
+    ".storybook/preview.tsx"
   ],
   "references": [
     {
