+++ libs/component-library/package.json
@@ -40,6 +40,7 @@
     "date-fns": "4.1.0",
     "lucide-react": "0.525.0",
     "next-themes": "0.4.6",
+    "posthog-js": "1.255.1",
     "react": "19.1.0",
     "react-day-picker": "8.10.1",
     "sonner": "2.0.5",
