+++ libs/component-library/src/storybook.ts
@@ -1,3 +1,7 @@
+import * as postHogMocks from "./storybook/providers/posthog.mock";
+
 export * from "./storybook/preview-config";
 export * from "./storybook/j5-sb-theme";
 export * from "./storybook/manager";
+
+export { postHogMocks };
