+++ libs/component-library/src/storybook/preview-config.tsx
@@ -1,6 +1,7 @@
 import type { Preview as PreviewVite } from "@storybook/react-vite";
 import { j5Theme } from "./j5-sb-theme";
-// import "../styles/globals.css";
+import { MockPostHogProvider } from "./providers/posthog.mock";
+import { TooltipProvider } from "../components/tooltip/tooltip";
 
 export const previewConfigVite = {
   parameters: {
@@ -16,4 +17,13 @@ export const previewConfigVite = {
   initialGlobals: {
     backgrounds: { value: "j5" },
   },
+  decorators: [
+    (Story) => (
+      <MockPostHogProvider>
+        <TooltipProvider>
+          <Story />
+        </TooltipProvider>
+      </MockPostHogProvider>
+    ),
+  ],
 } satisfies PreviewVite;
