+++ package.json
@@ -41,6 +41,7 @@
     "@nx/web": "21.2.1",
     "@playwright/test": "1.53.1",
     "@storybook/nextjs": "9.0.14",
+    "@storybook/nextjs-vite": "9.0.14",
     "@storybook/react-vite": "9.0.14",
     "@storybook/test-runner": "0.23.0",
     "@swc-node/register": "1.10.10",
